moved getDIRACGOCDictionary from SitesDIRACGOCDBmapping to ConfigurationSystem/Client/Utilities
